data-etl/requirements.txt
PyGithub==2.1.1
openai==1.54.0
httpx==0.27.2
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0
backoff==2.2.1
tqdm==4.66.1
pydantic==2.5.3
